+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/Casts/Tabs/Interrupts.lua
@@ -12,7 +12,7 @@ local function getIterruptedSpellMenu(rows, index, cols)
 	local spellId = rows[index].cols[1].value;
 	local report = MyDungeonsBook:InterruptsFrame_SpellInterruptedReport_Create(rows[index], cols);
 	return {
-		MyDungeonsBook:WowHead_Menu_Spell(spellId),
+		MyDungeonsBook:WowHead_Menu_SpellComplex(spellId),
 		MyDungeonsBook:Report_Menu(report)
 	};
 end