+++ UI/ChallengeDetails/Tabs/Mechanics/Tabs/EffectsAndAuras/Tabs/Dispels.lua
@@ -12,7 +12,7 @@ local function getDispellSpellMenu(rows, index, cols)
 	local spellId = rows[index].cols[1].value;
 	local report = MyDungeonsBook:DispelsFrame_SpellDispelledReport_Create(rows[index], cols);
 	return {
-		MyDungeonsBook:WowHead_Menu_Spell(spellId),
+		MyDungeonsBook:WowHead_Menu_SpellComplex(spellId),
 		MyDungeonsBook:Report_Menu(report)
 	};
 end